Fix various compiler warnings
Specifically,
-Wunused-variable
-Wsign-compare warnings        frequent int/unsigned int mismatches
-Woverloaded-virtual           Decoder::Module::Init() overload
-Wunused-but-set-variable      mostly leftovers from debugging
-Wformat-zero-length           sprintf(data,"")
-Wtype-limits                  testing unsigned for < 0
-Wmaybe-uninitialized          SBSGRINCH::Decode theHit (dubious indeed)
-Wextra                        Initialize base class in copy ctr

+++ SBSBBShower.cxx
@@ -178,7 +178,7 @@ void SBSBBShower::MakeCluster(Int_t nblk_size)
 
 void SBSBBShower::AddToCluster(Int_t nc,SBSElement* blk) 
 {
-  if (nc < fClusters.size()) fClusters[nc]->AddElement(blk);
+  if (nc < (int)fClusters.size()) fClusters[nc]->AddElement(blk);
 }
 
 void SBSBBShower::MakeMainCluster()
